social_media_api/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-authtoken
gunicorn
psycopg2-binary
